requirements.txt
svg.path>=4.1
PyAutoGUI==0.9.53 